Add password rules for ecompanystore.com (#160)
As flagged by Taylor Swift: https://twitter.com/swiftonsecurity/status/1270082231877263360?s=21

This wasn't doesn't allow a `?` to be the first character, so I omit it from the special characters set to ensure
it's never used as the first character. A future enhancement to the passwordRules language may accommodate this
constraint better.