Add password rule for lsacsso.b2clogin.com
Law school admission council password requirements:
> 8-16 characters, containing 3 out of 4 of the following: Lowercase
> characters, uppercase characters, digits (0-9), and one or more of the
> following symbols: @ # $ % ^ & * - _ ! . ?

The js source looks to use the following regex to perform this
validation:
> ( (?=.*[a-z])(?=.*[A-Z])(?=.*\d)|
>   (?=.*[a-z])(?=.*[A-Z])(?=.*[^A-Za-z0-9])|
>   (?=.*[a-z])(?=.*\d)(?=.*[^A-Za-z0-9])|
>   (?=.*[A-Z])(?=.*\d)(?=.*[^A-Za-z0-9])
> ) ([A-Za-z\d@#$%^&*\-_+=[\]{}|\\:',?/`~"();!]|\.(?!@)){8,16}

There are discrepancies between the regex and the text description of
the requirements. The regex allows more special characters but disallows
a "@" following a "." for some reason. The rule added by this commit
uses only the characters in the textual description, minus the "." (to
remove the possibility of a ".@" sequence).