Introduce the new "shared credentials" format
This patch implements the format described in https://github.com/apple/password-manager-resources/issues/464, with relatively
low effort to verify whether associations are current or historical. A few entries are made current, but most will start as
historical. (They can be moved out in subsequent pull requests.)

Then, this patch adds a script to take the two new files and produce a version of websites-with-shared-credential-backends.json
with their data.

Next, I clean up our GitHub workflow that reports whether files are sorted to track the two new files.

Finally, this patch runs the newly-added script to re-generate websites-with-shared-credential-backends.json from
shared-credentials.json and shared-credentials-historical.json.

We'll leave websites-with-shared-credential-backends.json in the tree for now.