Add instructions for regenerating websites-with-shared-credential-backends.json
quirks/websites-with-shared-credential-backends.json contains a lower fidelity version of
the data in quirks/shared-credentials.json and quirks/shared-credentials-historical.json.
It must be regenerated using tools/convert-shared-credential-to-legacy-format.rb whenever
those files are changed.

This information was not captuted in any of the current documentation, therefore
CONTRIBUTING.md has been updated to include it. The workflows we run on each PR have been
updated to verify that quirks/websites-with-shared-credential-backends.json is up-to-date.