Remove fromDomainsAreObsoleted from the overstock.com shared-credentials entry
`fromDomainsAreObsoleted` being set to `true` means that logins are no longer served from overstock.com. When I navigate to overstock.com's login page, it says to use a Bed Bath & Beyond account, but the page is still served from overstock.com. So, I think this flag should not be set on this rule.